Trait openzl_crypto::accumulator::Accumulator

pub trait Accumulator: Types {
    type Model: Model<Item = Self::Item, Witness = Self::Witness, Output = Self::Output> + ?Sized;

    fn model(&self) -> &Self::Model;
    fn insert(&mut self, item: &Self::Item) -> bool;
    fn prove(&self, item: &Self::Item) -> Option<MembershipProof<Self::Model>>;

    fn contains(&self, item: &Self::Item) -> bool { ... }
}

Accumulator

Required Associated Types

type Model: Model<Item = Self::Item, Witness = Self::Witness, Output = Self::Output> + ?Sized

Model Type

Required Methods

fn model(&self) -> &Self::Model

Returns the model associated with self.

fn insert(&mut self, item: &Self::Item) -> bool

Inserts item into self with the guarantee that self can later return a valid membership proof for item with a call to prove. This method returns false if the maximum capacity of the accumulator would be exceeded by inserting item.

fn prove(&self, item: &Self::Item) -> Option<MembershipProof<Self::Model>>

Returns a membership proof for item if it is contained in self.

Provided Methods

fn contains(&self, item: &Self::Item) -> bool

Returns true if item is stored in self.

Implementation Note

This method must at least return true for item whenever a valid proof of membership exists. It may return true in other cases when self knows that it has item stored but cannot return a proof for it, like in the case of OptimizedAccumulator implementations. In other words, this method is allowed to return false negatives, but not false positives.