Trait openzl_crypto::accumulator::AssertValidVerification

pub trait AssertValidVerification<COM = ()>: Model<COM> {
    fn assert_valid(
        &self,
        item: &Self::Item,
        witness: &Self::Witness,
        output: &Self::Output,
        compiler: &mut COM
    );
}

Accumulator Membership Model Validity Assertion

For situations where we just want to assert validity of the membership proof, we can use this trait as an optimization path for Model::verify. See assert_valid for more details.

Required Methods

fn assert_valid(
&self,
item: &Self::Item,
witness: &Self::Witness,
output: &Self::Output,
compiler: &mut COM
)

Asserts that the verification of the storage of item in the known accumulator is valid.

Optimization

In compilers where assertions for more complex statements other than booleans being true, this function can provide an optimization path to reduce the cost of assertion.