Trait openzl_crypto::accumulator::OptimizedAccumulator

pub trait OptimizedAccumulator: Accumulator {
    fn insert_nonprovable(&mut self, item: &Self::Item) -> bool { ... }
    fn remove_proof(&mut self, item: &Self::Item) -> bool { ... }
}

Optimized Accumulator

Provided Methods

fn insert_nonprovable(&mut self, item: &Self::Item) -> bool

Inserts item into self without the guarantee that self with be able to return a proof of membership for item. This method returns false if the maximum capacity of the accumulator would be exceeded by inserting item.

Implementation Note

By default, this method uses insert to store item in self. Since this method may insert items which will never have membership proofs, the contains method is allowed to return false for those items if necessary. In other words, the contains method is allowed to return false negatives, but not false positives.

fn remove_proof(&mut self, item: &Self::Item) -> bool

Removes the witnesses to the membership of item in self. The resulting state of the accumulator after removing a proof should be the same as if the item had been inserted into the accumulator with insert_nonprovable. This method returns true if the item was successfully demoted to non-provable.

Implementation Note

By default, this method does nothing and returns false. Implementations of this method may fail arbitrarily, and should only successfully remove a proof if the implementation is efficient enough. Space and time tradeoffs should be studied to determine the usefulness of this method.