Trait openzl_crypto::constraint::measure::Measure

pub trait Measure: Count<Constant> + Count<Public> + Count<Secret> {
    fn constraint_count(&self) -> usize;

    fn measure(&self) -> Size { ... }
    fn after<T, F>(&mut self, measurement: &mut Size, f: F) -> T
    where
        F: FnOnce(&mut Self) -> T,
    { ... }
    fn after_ignore<T, F>(&mut self, f: F) -> Size
    where
        F: FnOnce(&mut Self) -> T,
    { ... }
}

Constraint System Measurement

Required Methods

fn constraint_count(&self) -> usize

Returns the number of constraints stored in self.

Provided Methods

fn measure(&self) -> Size

Returns a Size with the number of constraints and variables of each kind.

fn after<T, F>(&mut self, measurement: &mut Size, f: F) -> Twhere
F: FnOnce(&mut Self) -> T,

Performs a measurement after running f on self, adding the result to measurement.

fn after_ignore<T, F>(&mut self, f: F) -> Sizewhere
F: FnOnce(&mut Self) -> T,

Performs a measurement after running f on self, ignoring the resulting value, returning the measurement only.