Trait openzl_crypto::constraint::ProofSystem

pub trait ProofSystem {
    type Compiler;
    type PublicParameters;
    type ProvingContext;
    type VerifyingContext;
    type Input: Default;
    type Proof;
    type Error;

    fn context_compiler() -> Self::Compiler;
    fn proof_compiler() -> Self::Compiler;
    fn compile<R>(
        public_parameters: &Self::PublicParameters,
        compiler: Self::Compiler,
        rng: &mut R
    ) -> Result<(Self::ProvingContext, Self::VerifyingContext), Self::Error>
    where
        R: CryptoRng + RngCore + ?Sized;
    fn prove<R>(
        context: &Self::ProvingContext,
        compiler: Self::Compiler,
        rng: &mut R
    ) -> Result<Self::Proof, Self::Error>
    where
        R: CryptoRng + RngCore + ?Sized;
    fn verify(
        context: &Self::VerifyingContext,
        input: &Self::Input,
        proof: &Self::Proof
    ) -> Result<bool, Self::Error>;
}

Proof System

Required Associated Types

type Compiler

Context Compiler

type PublicParameters

Public Parameters Type

type ProvingContext

Proving Context Type

type VerifyingContext

Verifying Context Type

type Input: Default

Verification Input Type

type Proof

Proof Type

type Error

Error Type

Required Methods

fn context_compiler() -> Self::Compiler

Returns a compiler which is setup to build proving and verifying contexts.

fn proof_compiler() -> Self::Compiler

Returns a compiler which is setup to build a proof.

fn compile<R>(
public_parameters: &Self::PublicParameters,
compiler: Self::Compiler,
rng: &mut R
) -> Result<(Self::ProvingContext, Self::VerifyingContext), Self::Error>where
R: CryptoRng + RngCore + ?Sized,

Returns proving and verifying contexts for the constraints contained in compiler using public_parameters.

fn prove<R>(
context: &Self::ProvingContext,
compiler: Self::Compiler,
rng: &mut R
) -> Result<Self::Proof, Self::Error>where
R: CryptoRng + RngCore + ?Sized,

Returns a proof that the constraint system encoded in compiler is consistent with the proving context.

fn verify(
context: &Self::VerifyingContext,
input: &Self::Input,
proof: &Self::Proof
) -> Result<bool, Self::Error>

Verifies that a proof generated from this proof system is valid.