Trait openzl_crypto::encryption::Derive

pub trait Derive<COM = ()>: DecryptionKeyType + EncryptionKeyType {
    fn derive(
        &self,
        decryption_key: &Self::DecryptionKey,
        compiler: &mut COM
    ) -> Self::EncryptionKey;
}

Encryption Key Derivation

For protocols that can derive the EncryptionKey from the DecryptionKey, this trait can be used to specify the derivation algorithm. The DecryptionKey should be kept secret relative to the EncryptionKey so if they are the same key, then some key exchange protocol should be used to derive the DecryptionKey as some shared secret. See hybrid for more.

Required Methods

fn derive(
&self,
decryption_key: &Self::DecryptionKey,
compiler: &mut COM
) -> Self::EncryptionKey

Derives an EncryptionKey from decryption_key.