Trait openzl_crypto::encryption::EncryptionKeyType

pub trait EncryptionKeyType {
    type EncryptionKey;
}

Encryption Key

The encryption key is the information required to produce a valid ciphertext that is targeted towards some DecryptionKey. In the case when the decryption key is linked by some computable protocol to the encryption key, Derive should be implemented to facilitate this derivation.

Required Associated Types

type EncryptionKey

Component Type