Module openzl_crypto::merkle_tree

Available on crate feature alloc only.

Merkle Trees and Forests

Modules

forest

Merkle Forests

fork

Merkle Tree Forks

full

Full Merkle Tree Storage

inner_tree

Inner Digest Tree

node 🔒

Merkle Tree Node Abstractions

partial

Partial Merkle Tree Storage

path

Merkle Tree Paths

single_path

Single Path Merkle Tree Storage

testtest

Testing Framework

tree 🔒

Merkle Tree Abstractions

Structs

Config

Configuration Structure

CurrentPath

Merkle Tree Current Path

IdentityLeafHash

Identity Leaf Hash

MerkleTree

Merkle Tree

Node

Node Index

NodeParents

Node Parent Iterator

Parameters

Merkle Tree Parameters

Path

Merkle Tree Path

Enums

ParameterDecodeError

Parameter Decode Error

Parity

Parity of a Subtree

PathError

Path Error

Traits

Configuration

Merkle Tree Configuration

HashConfiguration

Merkle Tree Hash Configuration

InnerHash

Merkle Tree Inner Hash

LeafHash

Merkle Tree Leaf Hash

Tree

Merkle Tree Structure

WithProofs

Merkle Tree Membership Proof Mixin

Functions

capacity

Returns the capacity of the merkle tree with the given C::HEIGHT parameter.

path_length

Returns the path length of the merkle tree with the given C::HEIGHT parameter.

Type Definitions

InnerDigest

Inner Hash Digest Type

InnerHashParameters

Inner Hash Parameters Type

Leaf

Leaf Type

LeafDigest

Leaf Hash Digest Type

LeafHashParameters

Leaf Hash Parameters Type

Root

Merkle Tree Root