Trait openzl_crypto::merkle_tree::forest::Forest

pub trait Forest<C>where
    C: Configuration + ?Sized,{
    type Tree: Tree<C>;

    fn new(parameters: &Parameters<C>) -> Self;
    fn len(&self) -> usize;
    fn capacity(&self) -> usize;
    fn get(&self, index: C::Index) -> &Self::Tree;
    fn get_mut(&mut self, index: C::Index) -> &mut Self::Tree;

    fn is_empty(&self) -> bool { ... }
    fn get_tree(&self, leaf: &Leaf<C>) -> &Self::Tree { ... }
    fn get_tree_mut(&mut self, leaf: &Leaf<C>) -> &mut Self::Tree { ... }
}

Available on crate feature alloc only.

Merkle Forest Structure

Required Associated Types

type Tree: Tree<C>

Tree Type

Required Methods

fn new(parameters: &Parameters<C>) -> Self

Builds a new empty merkle forest.

fn len(&self) -> usize

Returns the number of leaves in self.

fn capacity(&self) -> usize

Returns the number of leaves that can be stored in self.

fn get(&self, index: C::Index) -> &Self::Tree

Returns a shared reference to the tree at the given index.

Panics

This method is allowed to panic if index is out-of-bounds.

fn get_mut(&mut self, index: C::Index) -> &mut Self::Tree

Returns a mutable reference to the tree at the given index.

Panics

This method is allowed to panic if index is out-of-bounds.

Provided Methods

fn is_empty(&self) -> bool

Returns true if the length of self is zero.

fn get_tree(&self, leaf: &Leaf<C>) -> &Self::Tree

Returns a shared reference to the tree which leaf corresponds with.

Implementation Note

This method is an optimization path for computing tree_index followed by get.

fn get_tree_mut(&mut self, leaf: &Leaf<C>) -> &mut Self::Tree

Returns a mutable reference to the tree which leaf corresponds with.

Implementation Note

This method is an optimization path for computing tree_index followed by get_mut.