Struct openzl_crypto::merkle_tree::fork::Fork

pub struct Fork<C, T, P = SingleThreaded, M = BTreeMap<C>>where
    C: Configuration + ?Sized,
    T: Tree<C>,
    P: PointerFamily<T>,
    M: Default + InnerMap<C>,{
    base: P::Weak,
    branch: Branch<C, M>,
}

Available on crate feature alloc only.

Merkle Tree Fork

Fields
base: P::Weak

Base Merkle Tree

branch: Branch<C, M>

Branch Data

Implementations

impl<C, T, P, M> Fork<C, T, P, M>where
C: Configuration + ?Sized,
T: Tree<C>,
P: PointerFamily<T>,
M: Default + InnerMap<C>,

fn build(trunk: &Trunk<C, T, P>, branch: Branch<C, M>) -> Self

Builds a new Fork off of trunk with the given base_contribution and branch.

pub fn new(parameters: &Parameters<C>, trunk: &Trunk<C, T, P>) -> Selfwhere
LeafDigest<C>: Clone + Default,
InnerDigest<C>: Default,

Builds a new Fork from trunk.

pub fn with_leaves(
parameters: &Parameters<C>,
trunk: &Trunk<C, T, P>,
leaf_digests: Vec<LeafDigest<C>>
) -> Option<Self>where
LeafDigest<C>: Clone + Default,
InnerDigest<C>: Default,

Builds a new Fork from trunk extended by leaf_digests, returning None if appending leaf_digests would exceed the capacity of the trunk.

pub fn attach(
&mut self,
parameters: &Parameters<C>,
trunk: &Trunk<C, T, P>
) -> boolwhere
LeafDigest<C>: Clone + Default,
InnerDigest<C>: Default,

Tries to attach this fork to a new trunk, returning false if self has too many leaves to fit in trunk.

pub fn is_attached(&self) -> bool

Returns true if this fork is attached to some Trunk.

pub fn is_attached_to(&self, trunk: &Trunk<C, T, P>) -> bool

Returns true if this fork is attached to trunk.

fn get_attached_base(&self, trunk: &Trunk<C, T, P>) -> Option<P::Strong>

Returns the attached base tree if self is attached to trunk.

fn check_attachment(&self) -> Option<()>

Checks if self is still attached to trunk.

pub fn len(&self) -> usize

Computes the length of this fork of the tree.

pub fn is_empty(&self) -> bool

Returns true if this fork is empty.

pub fn root(&self) -> &InnerDigest<C>

Returns the current root of this fork.

pub fn leaf_digest(&self, index: usize) -> Option<&LeafDigest<C>>

Returns the leaf digest at the given index in the tree.

pub fn position(&self, leaf_digest: &LeafDigest<C>) -> Option<usize>where
T: WithProofs<C>,
LeafDigest<C>: PartialEq,

Returns the position of leaf_digest in the tree.

pub fn current_leaf(&self) -> Option<&LeafDigest<C>>

Returns the current (right-most) leaf of the tree.

pub fn current_path(&self) -> CurrentPath<C>where
LeafDigest<C>: Clone + Default,
InnerDigest<C>: Clone + Default + PartialEq,

Returns the current (right-most) path of the tree.

pub fn path(
&self,
parameters: &Parameters<C>,
index: usize
) -> Result<Path<C>, PathError>where
T: WithProofs<C>,
LeafDigest<C>: Clone + Default,
InnerDigest<C>: Clone,

Returns the path at the given index in the tree.

pub fn push(&mut self, parameters: &Parameters<C>, leaf: &Leaf<C>) -> Option<bool>where
LeafDigest<C>: Default,

Appends a new leaf onto this fork.

Returns None if this fork has been detached from its trunk. Use attach to re-associate a trunk to this fork.

pub fn maybe_push_digest<F>(
&mut self,
parameters: &Parameters<C>,
leaf_digest: F
) -> Option<bool>where
F: FnOnce() -> Option<LeafDigest<C>>,
LeafDigest<C>: Default,

Appends a new leaf_digest onto this fork.

Returns None if this fork has been detached from its trunk. Use attach to re-associate a trunk to this fork.