Struct openzl_crypto::merkle_tree::inner_tree::InnerTree

pub struct InnerTree<C, M = BTreeMap<C>, S = Sentinel<C>>where
    C: Configuration + ?Sized,
    M: InnerMap<C>,
    S: SentinelSource<C>,{
    map: M,
    sentinel_source: S,
    __: PhantomData<C>,
}

Available on crate feature alloc only.

Inner Tree

Tree data-structure for storing the inner digests of a merkle tree.

Implementation Note

This type intentionally lacks an implementation of Tree, especially since it does not store leaf digests. Instead, Full should be used whenever a tree with the capability to request arbitrary inner nodes is needed.

Fields
map: M

Inner Digest Map

Coordinates

Locations in the tree are indexed by two coordinates depth and index. The depth of a node is given by its layer in the tree starting from depth := -1 at the root increasing downwards towards the leaves. The index of a node is its position from left to right along a layer in the tree. See InnerNode for more details.

sentinel_source: S

Sentinel Source

The background tree for sentinel values.

__: PhantomData<C>

Type Parameter Marker

Implementations

impl<C, M, S> InnerTree<C, M, S>where
C: Configuration + ?Sized,
M: InnerMap<C>,
S: SentinelSource<C>,

pub fn new() -> Selfwhere
M: Default,
S: Default,

Builds a new InnerTree.

pub fn with_map(map: M) -> Selfwhere
S: Default,

Builds a new InnerTree with the given inner map.

pub fn with_sentinel(sentinel_source: S) -> Selfwhere
M: Default,

Builds a new InnerTree with the given sentinel_source.

pub fn with_map_and_sentinel(map: M, sentinel_source: S) -> Self

Builds a new InnerTree with the given map and sentinel_source.

pub fn get(&self, node: InnerNode) -> Option<&InnerDigest<C>>

Tries to get the inner digest at node, returning None if the inner digest is missing.

pub fn get_or_sentinel(&self, node: InnerNode) -> &InnerDigest<C>

Returns the inner digest at node or a sentinel value if the inner digest is missing.

pub fn map_get(&self, index: usize) -> Option<&InnerDigest<C>>

Tries to return the inner digest at index, returning None if the inner digest is missing.

pub fn map_get_or_sentinel(&self, index: usize) -> &InnerDigest<C>

Returns the inner digest at index or a sentinel value if the inner digest is missing.

pub fn root(&self) -> &InnerDigest<C>

Returns a reference to the root inner digest.

fn set_root(&mut self, root: InnerDigest<C>)

Sets the current root to root.

fn insert_and_get_pair(
&mut self,
node: InnerNode,
inner_digest: InnerDigest<C>
) -> (&InnerDigest<C>, &InnerDigest<C>)

Inserts the new inner_digest at node in the tree, and returns a reference to inner_digest and its sibling in the tree in parity order.

fn insert_and_join(
&mut self,
parameters: &Parameters<C>,
node: InnerNode,
inner_digest: InnerDigest<C>
) -> InnerDigest<C>

Inserts inner_digest into the tree at node and computes the join of inner_digest and its sibling in the tree, using the default value if its sibling is not stored in the tree.

fn compute_root(
&mut self,
parameters: &Parameters<C>,
leaf_index: Node,
base: InnerDigest<C>
) -> InnerDigest<C>

Computes the new root of the tree after inserting base which corresponds to the leaf at leaf_index.

pub fn insert(
&mut self,
parameters: &Parameters<C>,
leaf_index: Node,
base: InnerDigest<C>
)

Inserts the base inner digest corresponding to the leaf at leaf_index into the tree.

pub fn path_iter(&self, node: InnerNode) -> InnerTreePathIter<'_, C, M, S>

Computes the inner path starting from node.

pub fn path_iter_for_leaf(
&self,
leaf_index: Node
) -> InnerTreePathIter<'_, C, M, S>

Computes the inner path of the leaf given by leaf_index.

pub fn path(&self, leaf_index: Node) -> InnerPath<C>where
InnerDigest<C>: Clone,

Returns the path at leaf_index.

impl<C, M> InnerTree<C, M, Sentinel<C>>where
C: Configuration + ?Sized,
M: InnerMap<C>,
InnerDigest<C>: Clone + PartialEq,

pub fn current_path_unchecked(&self, leaf_index: Node) -> CurrentInnerPath<C>

Returns the path at leaf_index, assuming that leaf_index is the right-most index, so that the return value is a valid CurrentInnerPath.