Struct openzl_crypto::merkle_tree::partial::Partial

pub struct Partial<C, M = BTreeMap<C>>where
    C: Configuration + ?Sized,
    M: InnerMap<C>,{
    leaf_digests: Vec<LeafDigest<C>>,
    inner_digests: PartialInnerTree<C, M>,
}

Available on crate feature alloc only.

Partial Merkle Tree Backing Structure

Fields
leaf_digests: Vec<LeafDigest<C>>

Leaf Digests

inner_digests: PartialInnerTree<C, M>

Inner Digests

Implementations

impl<C, M> Partial<C, M>where
C: Configuration + ?Sized,
M: InnerMap<C>,

pub fn new_unchecked(
leaf_digests: Vec<LeafDigest<C>>,
inner_digests: PartialInnerTree<C, M>
) -> Self

Builds a new Partial without checking that leaf_digests and inner_digests form a consistent merkle tree.

pub fn leaf_digests(&self) -> &[LeafDigest<C>]

Returns the leaf digests currently stored in the merkle tree.

Note

Since this tree does not start its leaf nodes from the first possible index, indexing into this slice will not be the same as indexing into a slice from a full tree. For all other indexing, use the full indexing scheme.

pub fn into_leaves(self) -> Vec<LeafDigest<C>>

Returns the leaf digests stored in the tree, dropping the rest of the tree data.

Note

See the note at leaf_digests for more information on indexing this vector.

pub fn starting_leaf_node(&self) -> Node

Returns the starting leaf Node for this tree.

pub fn starting_leaf_index(&self) -> usize

Returns the starting leaf index for this tree.

pub fn len(&self) -> usize

Returns the number of leaves in this tree.

pub fn is_empty(&self) -> bool

Returns true if this tree is empty.

pub fn root(&self) -> &InnerDigest<C>

Returns a reference to the root inner digest.

pub fn leaf_digest(&self, index: usize) -> Option<&LeafDigest<C>>

Returns the leaf digest at the given index in the tree.

pub fn position(&self, leaf_digest: &LeafDigest<C>) -> Option<usize>where
LeafDigest<C>: PartialEq,

Returns the position of leaf_digest in the tree.

pub fn get_leaf_sibling(&self, index: Node) -> Option<&LeafDigest<C>>

Returns the sibling leaf node to index.

pub fn get_owned_leaf_sibling(&self, index: Node) -> LeafDigest<C>where
LeafDigest<C>: Clone + Default,

Returns an owned sibling leaf node to index.

pub fn current_leaf(&self) -> Option<&LeafDigest<C>>

Returns the current (right-most) leaf of the tree.

pub fn current_path(&self) -> CurrentPath<C>where
LeafDigest<C>: Clone + Default,
InnerDigest<C>: Clone + PartialEq,

Returns the current (right-most) path of the tree.

pub fn path_unchecked(&self, index: usize) -> Path<C>where
LeafDigest<C>: Clone + Default,
InnerDigest<C>: Clone,

Returns the path at index without bounds-checking on the index.

pub fn push_leaf_digest(
&mut self,
parameters: &Parameters<C>,
leaf_index: Node,
leaf_digest: LeafDigest<C>
)where
LeafDigest<C>: Default,

Appends a leaf_digest with index given by leaf_index into the tree.

pub fn push(&mut self, parameters: &Parameters<C>, leaf: &Leaf<C>) -> boolwhere
LeafDigest<C>: Default,

Appends a leaf to the tree using parameters.

pub fn maybe_push_digest<F>(
&mut self,
parameters: &Parameters<C>,
leaf_digest: F
) -> Option<bool>where
F: FnOnce() -> Option<LeafDigest<C>>,
LeafDigest<C>: Default,

Appends leaf_digest to the tree using parameters.