Struct openzl_crypto::merkle_tree::Parameters

pub struct Parameters<C, COM = ()>where
    C: HashConfiguration<COM> + ?Sized,{
    pub leaf: LeafHashParameters<C, COM>,
    pub inner: InnerHashParameters<C, COM>,
}

Available on crate feature alloc only.

Merkle Tree Parameters

Fields
leaf: LeafHashParameters<C, COM>

Leaf Hash Parameters

inner: InnerHashParameters<C, COM>

Inner Hash Parameters

Implementations

impl<C, COM> Parameters<C, COM>where
C: HashConfiguration<COM> + ?Sized,

pub fn new(
leaf: LeafHashParameters<C, COM>,
inner: InnerHashParameters<C, COM>
) -> Self

Builds a new Parameters from leaf and inner parameters.

pub fn digest_with(
&self,
leaf: &Leaf<C, COM>,
compiler: &mut COM
) -> LeafDigest<C, COM>

Computes the leaf digest of leaf using self.

pub fn join_with(
&self,
lhs: &InnerDigest<C, COM>,
rhs: &InnerDigest<C, COM>,
compiler: &mut COM
) -> InnerDigest<C, COM>

Combines two inner digests into a new inner digest using self.

pub fn join_leaves_with(
&self,
lhs: &LeafDigest<C, COM>,
rhs: &LeafDigest<C, COM>,
compiler: &mut COM
) -> InnerDigest<C, COM>

Combines two leaf digests into a new inner digest using self.

pub fn verify_path_with(
&self,
path: &PathVar<C, COM>,
root: &Root<C, COM>,
leaf: &Leaf<C, COM>,
compiler: &mut COM
) -> Bool<COM>where
C: Configuration<COM>,
COM: Has<bool>,
InnerDigest<C, COM>: ConditionalSwap<COM> + PartialEq<InnerDigest<C, COM>, COM>,
LeafDigest<C, COM>: ConditionalSwap<COM>,

Verify that path witnesses the fact that leaf is a member of a merkle tree with the given root.

impl<C> Parameters<C>where
C: HashConfiguration + ?Sized,

pub fn digest(&self, leaf: &Leaf<C>) -> LeafDigest<C>

Computes the leaf digest of leaf using self.

pub fn join(&self, lhs: &InnerDigest<C>, rhs: &InnerDigest<C>) -> InnerDigest<C>

Combines two inner digests into a new inner digest using self.

pub fn join_leaves(
&self,
lhs: &LeafDigest<C>,
rhs: &LeafDigest<C>
) -> InnerDigest<C>

Combines two leaf digests into a new inner digest using self.

pub fn verify_path(&self, path: &Path<C>, root: &Root<C>, leaf: &Leaf<C>) -> boolwhere
C: Configuration,
InnerDigest<C>: PartialEq,

Verify that path witnesses the fact that leaf is a member of a merkle tree with the given root.