Module openzl_crypto::merkle_tree::tree

Available on crate feature alloc only.

Merkle Tree Abstractions

Structs

Config

Configuration Structure

IdentityLeafHash

Identity Leaf Hash

MerkleTree

Merkle Tree

Parameters

Merkle Tree Parameters

Enums

ParameterDecodeError

Parameter Decode Error

PathError

Path Error

Traits

Configuration

Merkle Tree Configuration

HashConfiguration

Merkle Tree Hash Configuration

InnerHash

Merkle Tree Inner Hash

LeafHash

Merkle Tree Leaf Hash

Tree

Merkle Tree Structure

WithProofs

Merkle Tree Membership Proof Mixin

Functions

capacity

Returns the capacity of the merkle tree with the given C::HEIGHT parameter.

path_length

Returns the path length of the merkle tree with the given C::HEIGHT parameter.

Type Definitions

InnerDigest

Inner Hash Digest Type

InnerHashParameters

Inner Hash Parameters Type

Leaf

Leaf Type

LeafDigest

Leaf Hash Digest Type

LeafHashParameters

Leaf Hash Parameters Type

Root

Merkle Tree Root