Trait openzl_crypto::merkle_tree::tree::WithProofs

pub trait WithProofs<C>where
    C: Configuration + ?Sized,{
    fn leaf_digest(&self, index: usize) -> Option<&LeafDigest<C>>;
    fn position(&self, leaf_digest: &LeafDigest<C>) -> Option<usize>;
    fn maybe_push_provable_digest<F>(
        &mut self,
        parameters: &Parameters<C>,
        leaf_digest: F
    ) -> Option<bool>
    where
        F: FnOnce() -> Option<LeafDigest<C>>;
    fn path(
        &self,
        parameters: &Parameters<C>,
        index: usize
    ) -> Result<Path<C>, PathError>;

    fn contains(&self, leaf_digest: &LeafDigest<C>) -> bool { ... }
    fn push_provable_digest<F>(
        &mut self,
        parameters: &Parameters<C>,
        leaf_digest: F
    ) -> bool
    where
        F: FnOnce() -> LeafDigest<C>,
    { ... }
    fn push_provable(
        &mut self,
        parameters: &Parameters<C>,
        leaf: &Leaf<C>
    ) -> bool { ... }
    fn remove_path(&mut self, index: usize) -> bool { ... }
}

Available on crate feature alloc only.

Merkle Tree Membership Proof Mixin

Required Methods

fn leaf_digest(&self, index: usize) -> Option<&LeafDigest<C>>

Returns the leaf digest at the given index.

Implementation Note

This method is allowed to return None even if index is less than the current length of the tree. See position for more.

fn position(&self, leaf_digest: &LeafDigest<C>) -> Option<usize>

Returns the index of the leaf_digest if it is contained in self.

Implementation Note

This method is allowed to return None even if leaf_digest was inserted with a call to push_digest. This method need only return an index for leaves which are inserted with a call to push_provable.

fn maybe_push_provable_digest<F>(
&mut self,
parameters: &Parameters<C>,
leaf_digest: F
) -> Option<bool>where
F: FnOnce() -> Option<LeafDigest<C>>,

Checks if a leaf can be inserted into the tree and if it can, it runs leaf_digest to extract a leaf digest to insert, returning None if there was no leaf digest and returning Some(false) if the tree is at capacity. If this method is successful, the digest inserted will have an accompanying proof that can be returned by a call to the path method.

fn path(
&self,
parameters: &Parameters<C>,
index: usize
) -> Result<Path<C>, PathError>

Returns the path for the leaf stored at the given index if it exists.

Provided Methods

fn contains(&self, leaf_digest: &LeafDigest<C>) -> bool

Returns true if leaf_digest is provably stored in self.

See the position and push_provable methods for more.

fn push_provable_digest<F>(
&mut self,
parameters: &Parameters<C>,
leaf_digest: F
) -> boolwhere
F: FnOnce() -> LeafDigest<C>,

Appends leaf_digest to the end of the tree, retaining its path for later use with a call to the path method.

fn push_provable(&mut self, parameters: &Parameters<C>, leaf: &Leaf<C>) -> bool

Appends leaf to the end of the tree, retaining its path for later use with a call to the path method.

fn remove_path(&mut self, index: usize) -> bool

Removes a single path at the given index, returning true if it was removed.

Implementation Note

This method may return false for arbitrary inputs and is only an optimization path for removing unused memory.