Trait openzl_crypto::permutation::duplex::Setup

pub trait Setup<P, COM = ()>: Types<P, COM>where
    P: PseudorandomPermutation<COM>,{
    fn initialize(&self, compiler: &mut COM) -> P::Domain;
    fn setup(
        &self,
        key: &Self::Key,
        header: &Self::Header,
        compiler: &mut COM
    ) -> Vec<Self::SetupBlock>;
}

Available on crate feature alloc only.

Duplex Sponge Initialization and Setup

Required Methods

fn initialize(&self, compiler: &mut COM) -> P::Domain

Initializes the Sponge state for the beginning of the cipher.

fn setup(
&self,
key: &Self::Key,
header: &Self::Header,
compiler: &mut COM
) -> Vec<Self::SetupBlock>

Generates the starting input blocks for key and header data to be inserted into the cipher.