Trait openzl_crypto::poseidon::Specification

pub trait Specification<COM = ()>: Field<COM> + Constants + Sized {
    fn apply_sbox(point: &mut Self::Field, compiler: &mut COM);

    fn mds_matrix_multiply(
        mds_matrix: &[Self::ParameterField],
        state: &mut State<Self, COM>,
        compiler: &mut COM
    ) { ... }
    fn full_round(
        additive_keys_current_round: &[Self::ParameterField],
        mds_matrix: &[Self::ParameterField],
        state: &mut State<Self, COM>,
        compiler: &mut COM
    ) { ... }
    fn partial_round(
        additive_keys_current_round: &[Self::ParameterField],
        mds_matrix: &[Self::ParameterField],
        state: &mut State<Self, COM>,
        compiler: &mut COM
    ) { ... }
    fn permute_without_first_round(
        additive_round_keys: &[Self::ParameterField],
        mds_matrix: &[Self::ParameterField],
        state: &mut State<Self, COM>,
        compiler: &mut COM
    ) { ... }
    fn first_round_with_domain_tag_unchecked<const N: usize>(
        domain_tag: &Self::Field,
        additive_keys_current_round: &[Self::ParameterField],
        mds_matrix: &[Self::ParameterField],
        input: [&Self::Field; N],
        compiler: &mut COM
    ) -> State<Self, COM> { ... }
}

Available on crate feature alloc only.

Poseidon Permutation Specification

Implementation Note

This trait includes a blanket implementation of mds_matrix_multiply that may not be optimal for all choices of COM. In particular, Plonk-like arithmetizations should implement mds_matrix_multiply in a way that minimizes the cost of linear combinations.

Required Methods

fn apply_sbox(point: &mut Self::Field, compiler: &mut COM)

Applies the S-BOX to point.

Provided Methods

fn mds_matrix_multiply(
mds_matrix: &[Self::ParameterField],
state: &mut State<Self, COM>,
compiler: &mut COM
)

Computes the MDS matrix multiplication against the state.

The argument mds_matrix is assumed to be the flattening of a matrix of size Self::WIDTH * Self::WIDTH.

fn full_round(
additive_keys_current_round: &[Self::ParameterField],
mds_matrix: &[Self::ParameterField],
state: &mut State<Self, COM>,
compiler: &mut COM
)

Computes a full round on the internal permutation state.

fn partial_round(
additive_keys_current_round: &[Self::ParameterField],
mds_matrix: &[Self::ParameterField],
state: &mut State<Self, COM>,
compiler: &mut COM
)

Computes a partial round on the internal permutation state.

fn permute_without_first_round(
additive_round_keys: &[Self::ParameterField],
mds_matrix: &[Self::ParameterField],
state: &mut State<Self, COM>,
compiler: &mut COM
)

Computes the full permutation without the first round.

Unlike full_round and partial_round this takes the full array of additive round constants for the permutation.

fn first_round_with_domain_tag_unchecked<const N: usize>(
domain_tag: &Self::Field,
additive_keys_current_round: &[Self::ParameterField],
mds_matrix: &[Self::ParameterField],
input: [&Self::Field; N],
compiler: &mut COM
) -> State<Self, COM>

Computes the first round borrowing the input and domain_tag returning the State after the first round. This method does not check that N + 1 = S::WIDTH.