Trait openzl_crypto::signature::Verify

pub trait Verify<COM = ()>: MessageType + SignatureType + VerifyingKeyType {
    type Verification;

    fn verify(
        &self,
        verifying_key: &Self::VerifyingKey,
        message: &Self::Message,
        signature: &Self::Signature,
        compiler: &mut COM
    ) -> Self::Verification;
}

Signature Verification

Required Associated Types

type Verification

Verification Result Type

This type is typically either bool, a Result type, or a compiler variable representing either of those concrete types.

Required Methods

fn verify(
&self,
verifying_key: &Self::VerifyingKey,
message: &Self::Message,
signature: &Self::Signature,
compiler: &mut COM
) -> Self::Verification

Verifies that the signature of message was signed with the signing key deriving verifying_key.

For correctness of the signature, verifying_key should have come from a call to Derive::derive, performed by the signer.